[SPARK-46517][PS][TESTS][FOLLOWUPS] Reorganize `IndexingTest`: factor out `test_loc2d*`
### What changes were proposed in this pull request?
1, factor out `test_loc2d*`;
2, add the missing parity tests;

### Why are the changes needed?
test parity and testing parallelism

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44518 from zhengruifeng/ps_test_indexing_loc2d.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
